Change submodule url to HTTPS
Due to ssh requires host to add public key to Github server first before cloning the repository. This cause it not working well with Docker. This change is made so that there is no need to setup public key with Github server and allow Docker container to successfully clone submodule.